[refactor] #334 도메인 모델 패키지 이동
- presentation 레이어에 있던 InterestType, LoginType 모델을 domain 레이어로 이동
- 관련 import 경로 수정
