+++ domain/src/main/java/kr/techit/lion/domain/model/InterestType.kt
@@ -1,4 +1,4 @@
-package kr.techit.lion.presentation.login.model
+package kr.techit.lion.domain.model
 
 enum class  InterestType {
     Physical, Hear, Visual, Elderly, Child
