+++ domain/src/main/java/kr/techit/lion/domain/model/LoginType.kt
@@ -1,0 +1,5 @@
+package kr.techit.lion.domain.model
+
+enum class LoginType {
+    NAVER, KAKAO
+}
