--- presentation/src/main/java/kr/techit/lion/presentation/login/model/LoginType.kt
@@ -1,5 +1,0 @@
-package kr.techit.lion.presentation.login.model
-
-enum class LoginType {
-    NAVER, KAKAO
-}
